+++ PrivateFrameworks/VectorKit.framework/VKTileProvider.h
@@ -31,9 +31,9 @@
         struct Matrix<double, 4, 4> { 
             double _e[16]; 
         } orientation; 
-        float aspectRatio; 
-        float verticalFieldOfView; 
-        float horizontalOffset; 
+        double aspectRatio; 
+        double verticalFieldOfView; 
+        double horizontalOffset; 
     }  _lastCameraState;
     struct CGSize { 
         double width;
